+++ packages/sdk/browser/contract-tests/entity/src/TestHarnessWebSocket.ts
@@ -43,6 +43,7 @@ export default class TestHarnessWebSocket {
             'strongly-typed',
             'client-prereq-events',
             'client-per-context-summaries',
+            'track-hooks',
           ];
 
           break;