+++ packages/sdk/browser/src/common.ts
@@ -32,6 +32,12 @@ export type {
   LDLogLevel,
   LDMultiKindContext,
   LDSingleKindContext,
+  TrackSeriesContext,
+  LDPluginBase,
+  LDPluginEnvironmentMetadata,
+  LDPluginSdkMetadata,
+  LDPluginApplicationMetadata,
+  LDPluginMetadata,
 } from '@launchdarkly/js-client-sdk-common';
 
 /**