+++ packages/sdk/react-native/__tests__/MobileDataManager.test.ts
@@ -87,11 +87,13 @@ describe('given a MobileDataManager with mocked dependencies', () => {
       trackEventModifier: (event) => event,
       hooks: [],
       inspectors: [],
+      getImplementationHooks: () => [],
+      credentialType: 'mobileKey',
     };
     const mockedFetch = mockFetch('{"flagA": true}', 200);
     platform = {
       crypto: new PlatformCrypto(),
-      info: new PlatformInfo(config.logger),
+      info: new PlatformInfo(config.logger, {}),
       requests: {
         createEventSource: jest.fn((streamUri: string = '', options: any = {}) => ({
           streamUri,