+++ packages/shared/common/src/api/index.ts
@@ -4,3 +4,4 @@ export * from './data';
 export * from './options';
 export * from './platform';
 export * as subsystem from './subsystem';
+export * from './integrations';